requirements.txt
pillow==5.3.0
ConfigParser==3.5.0